Split the repl modules into smaller files
